DentalAI-Intelligent-Oral-Health-Analyzer-Consultation-Platform-main/DentalAI-Intelligent-Oral-Health-Analyzer-Consultation-Platform-main/backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
bcrypt==5.0.0
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
ecdsa==0.19.1
fastapi==0.123.0
filelock==3.20.0
fonttools==4.61.0
fsspec==2025.10.0
greenlet==3.2.4
h11==0.16.0
httptools==0.7.1
idna==3.11
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.7
mpmath==1.3.0
networkx==3.6
numpy==2.3.5
packaging==25.0
passlib==1.7.4
pillow==12.0.0
psycopg2-binary==2.9.11
pyasn1==0.6.1
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.20
PyYAML==6.0.3
rsa==4.9.1
scikit-learn==1.7.2
scipy==1.16.3
six==1.17.0
SQLAlchemy==2.0.44
sqlmodel==0.0.27
starlette==0.50.0
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.7.1+cu118
torchaudio==2.7.1+cu118
torchvision==0.22.1+cu118
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1
